requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
google-generativeai
sentence-transformers
faiss-cpu
numpy
python-dotenv
Jinja2
